Move FITSDiff open-files test to existing test module
